[DOC] improved README structure to improve user journey (#141)
This PR improves the structure of the README to be in line with the
common user persona of someone wanting to understand what `tsbootstrap`
offers.

As a consequence, this is mainly a reordering:

1. basic usage vignette
2. installation
3. structure
4. roadmap
5. contributing

Further, the setup/instruction sections have been shortened and
streamlined significantly. For instance, "distractions" such as praising
`uv` are mostly removed (I also like it, but it makes the exposition
more meandering...)

---------

Co-authored-by: Sankalp Gilda <sankalp.gilda@gmail.com>